Add stories, split view component, add delete error handling
