+++ src/components/annuities/index.ts
@@ -16,5 +16,7 @@ export type {
   EditAnnuityDialogProps,
   EditAnnuityInput,
 } from "./EditAnnuityDialog";
-export { EditAnnuityDialog, EditAnnuityDialogView } from "./EditAnnuityDialog";
+export { EditAnnuityDialog } from "./EditAnnuityDialog";
+export type { EditAnnuityDialogViewProps } from "./EditAnnuityDialogView";
+export { EditAnnuityDialogView } from "./EditAnnuityDialogView";
 export { NewAnnuityDialog } from "./NewAnnuityDialog";
